+++ frontend/vue.config.js
@@ -4,14 +4,17 @@ const CompressionPlugin = require('compression-webpack-plugin');
 const LicenseWebpackPlugin = require('license-webpack-plugin').LicenseWebpackPlugin;
 
 const defaultTheme = { 'theme-darkly': './src/themes/darkly.scss' };
-const allThemes = {
+const minThemes = {
     ...defaultTheme,
+    ...{ 'theme-flatly' : './src/themes/flatly.scss' }
+};
+const allThemes = {
+    ...minThemes,
     ...{
         'theme-basic'    : './src/themes/basic.scss',
         'theme-cerulean' : './src/themes/cerulean.scss',
         'theme-cosmo'    : './src/themes/cosmo.scss',
         'theme-cyborg'   : './src/themes/cyborg.scss',
-        'theme-flatly'   : './src/themes/flatly.scss',
         'theme-journal'  : './src/themes/journal.scss',
         'theme-litera'   : './src/themes/litera.scss',
         'theme-lumen'    : './src/themes/lumen.scss',
@@ -35,20 +38,23 @@ const allThemes = {
     }
 };
 
+// Set to true to be able to switch themes in dev mode, but CSS reloading will not work anymore.
+const devModeThemes = false;
+
+// Use "minThemes" to build only 2 themes, which is a lot faster.
+const useThemes = allThemes;
+
 module.exports = defineConfig(() => {
     const production = process.env.NODE_ENV === 'production';
     return {
         outputDir: path.resolve(__dirname, '../web/dist'),
         publicPath: production ? '/dist/' : '/',
-        css: {
-            extract: true, // necessary for themes in dev mode
-        },
+        css: { extract: production || devModeThemes },
         configureWebpack: config => {
             config.resolve = {
                 fallback: { 'querystring': require.resolve('querystring-es3') },
             };
-            config.entry = allThemes;
-            //config.entry = defaultTheme;
+            config.entry = production || devModeThemes ? useThemes : defaultTheme;
             config.entry.main = './src/main.js';
             if (production) {
                 config.plugins.push(new CompressionPlugin({
@@ -56,9 +62,7 @@ module.exports = defineConfig(() => {
                     threshold: 1,
                     compressionOptions: { level: 6 },
                 }));
-                config.plugins.push(new LicenseWebpackPlugin({
-                    perChunkOutput: false,
-                }));
+                config.plugins.push(new LicenseWebpackPlugin({ perChunkOutput: false }));
             }
         },
         chainWebpack: config => {
@@ -67,26 +71,24 @@ module.exports = defineConfig(() => {
                 .test(/datatables\.net.*\.js$/)
                 .use('imports-loader')
                 .loader('imports-loader')
-                .options({
-                    additionalCode: 'var define = false;', // Disable AMD
-                })
+                .options({ additionalCode: 'var define = false;', }) // Disable AMD
                 .end()
             config.plugin('html').tap(args => {
-                args[0].inject = false; // files are manually injected in index.html
+                if (production || devModeThemes) {
+                    args[0].inject = false; // Files are manually injected in index.html.
+                }
                 if (production) {
                     args[0].filename = path.resolve(__dirname, '../web/index.html');
                 }
                 return args;
             });
-            config.plugin('copy').tap(args => {
-                if (production) {
+            if (production) {
+                config.plugin('copy').tap(args => {
                     // This favicon.ico is only used for dev mode to prevent 404 errors.
                     // For production, it's already in web/favicon.ico.
                     args[0].patterns[0].globOptions.ignore.push(path.resolve(__dirname, 'public/favicon.ico'));
-                }
-                return args;
-            });
-            if (production) {
+                    return args;
+                });
                 config.plugin('progress').use(require('webpack/lib/ProgressPlugin'))
             }
         },